@@ -23,11 +23,11 @@
         }
     ],
     "require": {
-        "php": "^8.1",
-        "filament/filament": "^3.0",
+        "php": "^8.3",
+        "filament/filament": "^4.0",
         "pragmarx/google2fa": "^8.0",
         "bacon/bacon-qr-code": "^3.0",
-        "spatie/laravel-passkeys": "^1.0"   ,
+        "spatie/laravel-passkeys": "^1.0",
         "spatie/laravel-package-tools": "^1.15.0"
     },
     "require-dev": {
