@@ -2,14 +2,15 @@
 
 namespace Stephenjude\FilamentTwoFactorAuthentication\Livewire;
 
-use DanHarrin\LivewireRateLimiting\Exceptions\TooManyRequestsException;
 use Filament\Actions\Action;
 use Filament\Actions\Contracts\HasActions;
-use Filament\Forms\Components\Placeholder;
 use Filament\Forms\Components\TextInput;
 use Filament\Forms\Contracts\HasForms;
-use Filament\Forms\Form;
+use Filament\Infolists\Components\TextEntry;
+use Filament\Schemas\Components\Actions;
+use Filament\Schemas\Schema;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\HtmlString;
 use Livewire\Component;
 use Stephenjude\FilamentTwoFactorAuthentication\Actions\ConfirmTwoFactorAuthentication;
 use Stephenjude\FilamentTwoFactorAuthentication\Actions\DisableTwoFactorAuthentication;
@@ -27,9 +28,7 @@ class TwoFactorAuthentication extends Component implements HasActions, HasForms
 
     public ?string $redirectTo = null;
 
-    public bool $isConfirmingSetup = false;
-
-    public bool $showRecoveryCodes = false;
+    public bool $showSetupCode = false;
 
     public function mount(): void
     {
@@ -41,179 +40,197 @@ class TwoFactorAuthentication extends Component implements HasActions, HasForms
         return view('filament-two-factor-authentication::livewire.two-factor-authentication');
     }
 
-    public function confirmSetup(): void
+    public function setupTwoFactorAuthenticationForm(Schema $schema): Schema
     {
-        try {
-            $this->rateLimit(5);
-
-            $data = $this->form->getState();
-
-            app(ConfirmTwoFactorAuthentication::class)($this->getUser(), $data['code']);
-
-            $this->isConfirmingSetup = false;
-            $this->showRecoveryCodes = true;
-
-            if ($this->redirectTo) {
-                redirect()->to($this->redirectTo);
-            }
-        } catch (TooManyRequestsException $exception) {
-            $this->sendRateLimitedNotification($exception);
-
-            return;
-        }
-    }
-
-    public function confirmSetupAction(): Action
-    {
-        return Action::make('confirmSetup')
-            ->label(__('filament-two-factor-authentication::components.2fa.confirm'))
-            ->visible(fn () => $this->isConfirmingSetup)
-            ->submit('confirmSetup');
-    }
-
-    public function cancelSetupAction(): Action
-    {
-        return Action::make('cancelSetup')
-            ->label(__('filament-two-factor-authentication::components.2fa.cancel'))
-            ->outlined()
-            ->visible(fn () => $this->isConfirmingSetup)
-            ->action(function () {
-                try {
-                    $this->rateLimit(5);
-
-                    app(DisableTwoFactorAuthentication::class)($this->getUser());
-
-                    $this->isConfirmingSetup = false;
-                } catch (TooManyRequestsException $exception) {
-                    $this->sendRateLimitedNotification($exception);
-
-                    return;
-                }
-            });
-    }
-
-    protected function enableTwoFactorAuthenticationAction(): Action
-    {
-        return Action::make('enableTwoFactorAuthentication')
-            ->label(__('filament-two-factor-authentication::components.2fa.enable'))
-            ->visible(
-                fn () => ! $this->getUser()->hasEnabledTwoFactorAuthentication()
-            )->modalWidth('md')
-            ->modalSubmitActionLabel(__('filament-two-factor-authentication::components.2fa.confirm'))
-            ->form(function () {
-                if (! TwoFactorAuthenticationPlugin::get()->twoFactorSetupRequiresPassword()) {
-                    return null;
-                }
-
-                return [
-                    TextInput::make('confirmPassword')
-                        ->label(__('filament-two-factor-authentication::components.2fa.confirm_password'))
-                        ->password()
-                        ->revealable(filament()->arePasswordsRevealable())
-                        ->required()
-                        ->autocomplete('confirm-password')
-                        ->rules([
-                            fn () => function (string $attribute, $value, $fail) {
-                                if (! Hash::check($value, $this->getUser()->password)) {
-                                    $fail(__('filament-two-factor-authentication::components.2fa.wrong_password'));
-                                }
-                            },
-                        ]),
-                ];
-            })
-            ->action(function () {
-                try {
-                    $this->rateLimit(5);
-
-                    app(EnableTwoFactorAuthentication::class)($this->getUser());
-
-                    $this->isConfirmingSetup = true;
-                } catch (TooManyRequestsException $exception) {
-                    $this->sendRateLimitedNotification($exception);
-
-                    return;
-                }
-            });
-    }
-
-    public function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                Placeholder::make('setup_key')
-                    ->label(fn () => __(
-                        'filament-two-factor-authentication::components.2fa.setup_key',
-                        ['setup_key' => decrypt($this->getUser()->two_factor_secret)]
-                    )),
+        return $schema
+            ->live()
+            ->statePath('data')
+            ->model($this->getUser())
+            ->hidden(fn () => ! $this->showSetupCode)
+            ->components([
+                TextEntry::make('header')
+                    ->hiddenLabel()
+                    ->state(__('filament-two-factor-authentication::components.setup_confirmation.header')),
+                TextEntry::make('description')
+                    ->hiddenLabel()
+                    ->state(
+                        __('filament-two-factor-authentication::components.setup_confirmation.description')
+                    ),
+                TextEntry::make('notice')
+                    ->hiddenLabel()
+                    ->state(
+                        __('filament-two-factor-authentication::components.setup_confirmation.scan_qr_code')
+                    ),
+                TextEntry::make('qrcode')
+                    ->hiddenLabel()
+                    ->state(
+                        fn ($record) => $record->two_factor_secret ? new HtmlString($record->twoFactorQrCodeSvg()) : ''
+                    ),
+                TextEntry::make('setup_key')
+                    ->label(fn ($record) => __('filament-two-factor-authentication::components.2fa.setup_key', [
+                        'setup_key' => $record->two_factor_secret ? decrypt($record->two_factor_secret) : '',
+                    ])),
                 TextInput::make('code')
                     ->label(__('filament-two-factor-authentication::components.2fa.code'))
                     ->required(),
-            ])
-            ->statePath('data');
+                Actions::make([
+                    Action::make('confirm')
+                        ->label(__('filament-two-factor-authentication::components.2fa.confirm'))
+                        ->action(function ($record) {
+                            $data = $this->setupTwoFactorAuthenticationForm->getState();
+
+                            app(ConfirmTwoFactorAuthentication::class)($record, $data['code']);
+
+                            $this->showSetupCode = false;
+                        }),
+                    Action::make('cancel')
+                        ->label(__('filament-two-factor-authentication::components.2fa.cancel'))
+                        ->outlined()
+                        ->action(function ($record) {
+                            $this->showSetupCode = false;
+
+                            app(DisableTwoFactorAuthentication::class)($record);
+                        }),
+                ]),
+            ]);
     }
 
-    protected function disableTwoFactorAuthenticationAction(): Action
+    public function enableTwoFactorAuthenticationForm(Schema $schema): Schema
     {
-        return Action::make('disableTwoFactorAuthentication')
-            ->label(__('filament-two-factor-authentication::components.2fa.disable'))
-            ->color('danger')
-            ->visible(fn () => $this->getUser()->hasEnabledTwoFactorAuthentication())
-            ->modalWidth('md')
-            ->modalSubmitActionLabel(__('filament-two-factor-authentication::components.2fa.confirm'))
-            ->form(function () {
-                if (! TwoFactorAuthenticationPlugin::get()->twoFactorSetupRequiresPassword()) {
-                    return null;
-                }
+        return $schema
+            ->live()
+            ->statePath('data')
+            ->model($this->getUser())
+            ->hidden(fn ($record) => $record->hasEnabledTwoFactorAuthentication() || $this->showSetupCode)
+            ->components([
+                TextEntry::make('header')
+                    ->hiddenLabel()
+                    ->state(__('filament-two-factor-authentication::components.enable.header')),
+                TextEntry::make('description')
+                    ->hiddenLabel()
+                    ->state(__('filament-two-factor-authentication::components.enable.description')),
+                Actions::make([
+                    Action::make('enableTwoFactorAuthentication')
+                        ->modalWidth('md')
+                        ->label(__('filament-two-factor-authentication::components.2fa.enable'))
+                        ->modalSubmitActionLabel(__('filament-two-factor-authentication::components.2fa.confirm'))
+                        ->action(function ($record) {
+                            $this->showSetupCode = true;
 
-                return [
-                    TextInput::make('currentPassword')
-                        ->label(__('filament-two-factor-authentication::components.2fa.current_password'))
-                        ->password()
-                        ->revealable(filament()->arePasswordsRevealable())
-                        ->required()
-                        ->autocomplete('current-password')
-                        ->rules([
-                            fn () => function (string $attribute, $value, $fail) {
-                                if (! Hash::check($value, $this->getUser()->password)) {
-                                    $fail(__('filament-two-factor-authentication::components.2fa.wrong_password'));
-                                }
-                            },
-                        ]),
-                ];
-            })
-            ->action(fn () => app(DisableTwoFactorAuthentication::class)($this->getUser()));
+                            app(EnableTwoFactorAuthentication::class)($record);
+                        })
+                        ->schema(function () {
+                            if (! TwoFactorAuthenticationPlugin::get()->twoFactorSetupRequiresPassword()) {
+                                return null;
+                            }
+
+                            return [
+                                TextInput::make('confirmPassword')
+                                    ->label(__('filament-two-factor-authentication::components.2fa.confirm_password'))
+                                    ->password()
+                                    ->revealable(filament()->arePasswordsRevealable())
+                                    ->required()
+                                    ->autocomplete('confirm-password')
+                                    ->rules([
+                                        fn () => function (string $attribute, $value, $fail) {
+                                            if (! Hash::check($value, $this->getUser()->password)) {
+                                                $fail(
+                                                    __(
+                                                        'filament-two-factor-authentication::components.2fa.wrong_password'
+                                                    )
+                                                );
+                                            }
+                                        },
+                                    ]),
+                            ];
+                        }),
+                ]),
+            ]);
     }
 
-    protected function generateNewRecoveryCodesAction(): Action
+    public function disableTwoFactorAuthenticationForm(Schema $schema): Schema
     {
-        return Action::make('generateNewRecoveryCodes')
-            ->label(__('filament-two-factor-authentication::components.2fa.regenerate_recovery_codes'))
-            ->outlined()
-            ->visible(fn () => $this->getUser()->hasEnabledTwoFactorAuthentication())
-            ->requiresConfirmation(! TwoFactorAuthenticationPlugin::get()->twoFactorSetupRequiresPassword())
-            ->modalWidth('md')
-            ->modalSubmitActionLabel(__('filament-two-factor-authentication::components.2fa.confirm'))
-            ->form(function () {
-                if (! TwoFactorAuthenticationPlugin::get()->twoFactorSetupRequiresPassword()) {
-                    return null;
-                }
+        return $schema
+            ->live()
+            ->statePath('data')
+            ->model($this->getUser())
+            ->hidden(fn () => ! $this->getUser()->hasEnabledTwoFactorAuthentication())
+            ->components([
+                TextEntry::make('recoveryCode')
+                    ->hiddenLabel()
+                    ->listWithLineBreaks()
+                    ->copyable()
+                    ->state(
+                        fn () => $this->getUser()->hasEnabledTwoFactorAuthentication()
+                            ? $this->getUser()->recoveryCodes()
+                            : []
+                    ),
+                Actions::make([
+                    Action::make('generateNewRecoveryCodes')
+                        ->label(__('filament-two-factor-authentication::components.2fa.regenerate_recovery_codes'))
+                        ->outlined()
+                        ->requiresConfirmation(! TwoFactorAuthenticationPlugin::get()->twoFactorSetupRequiresPassword())
+                        ->modalWidth('md')
+                        ->modalSubmitActionLabel(__('filament-two-factor-authentication::components.2fa.confirm'))
+                        ->action(fn ($record) => app(GenerateNewRecoveryCodes::class)($record))
+                        ->schema(function () {
+                            if (! TwoFactorAuthenticationPlugin::get()->twoFactorSetupRequiresPassword()) {
+                                return null;
+                            }
 
-                return [
-                    TextInput::make('currentPassword')
-                        ->label(__('filament-two-factor-authentication::components.2fa.current_password'))
-                        ->password()
-                        ->revealable(filament()->arePasswordsRevealable())
-                        ->required()
-                        ->autocomplete('current-password')
-                        ->rules([
-                            fn () => function (string $attribute, $value, $fail) {
-                                if (! Hash::check($value, $this->getUser()->password)) {
-                                    $fail(__('filament-two-factor-authentication::components.2fa.wrong_password'));
-                                }
-                            },
-                        ]),
-                ];
-            })
-            ->action(fn () => app(GenerateNewRecoveryCodes::class)($this->getUser()), $this->showRecoveryCodes = true);
+                            return [
+                                TextInput::make('currentPassword')
+                                    ->label(__('filament-two-factor-authentication::components.2fa.current_password'))
+                                    ->password()
+                                    ->revealable(filament()->arePasswordsRevealable())
+                                    ->required()
+                                    ->autocomplete('current-password')
+                                    ->rules([
+                                        fn () => function (string $attribute, $value, $fail) {
+                                            if (! Hash::check($value, $this->getUser()->password)) {
+                                                $fail(
+                                                    __(
+                                                        'filament-two-factor-authentication::components.2fa.wrong_password'
+                                                    )
+                                                );
+                                            }
+                                        },
+                                    ]),
+                            ];
+                        }),
+                    Action::make('disableTwoFactorAuthentication')
+                        ->label(__('filament-two-factor-authentication::components.2fa.disable'))
+                        ->color('danger')
+                        ->modalWidth('md')
+                        ->modalSubmitActionLabel(__('filament-two-factor-authentication::components.2fa.confirm'))
+                        ->action(fn ($record) => app(DisableTwoFactorAuthentication::class)($record))
+                        ->requiresConfirmation()
+                        ->schema(function () {
+                            if (! TwoFactorAuthenticationPlugin::get()->twoFactorSetupRequiresPassword()) {
+                                return null;
+                            }
+
+                            return [
+                                TextInput::make('currentPassword')
+                                    ->label(__('filament-two-factor-authentication::components.2fa.current_password'))
+                                    ->password()
+                                    ->revealable(filament()->arePasswordsRevealable())
+                                    ->required()
+                                    ->autocomplete('current-password')
+                                    ->rules([
+                                        fn () => function (string $attribute, $value, $fail) {
+                                            if (! Hash::check($value, $this->getUser()->password)) {
+                                                $fail(
+                                                    __(
+                                                        'filament-two-factor-authentication::components.2fa.wrong_password'
+                                                    )
+                                                );
+                                            }
+                                        },
+                                    ]),
+                            ];
+                        }),
+                ]),
+            ]);
     }
 }
