@@ -8,9 +8,9 @@ use Exception;
 use Filament\Actions\Concerns\InteractsWithActions;
 use Filament\Facades\Filament;
 use Filament\Forms\Concerns\InteractsWithForms;
-use Filament\Models\Contracts\FilamentUser;
 use Filament\Notifications\Notification;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Foundation\Auth\User;
 
 trait Defaults
 {
@@ -18,7 +18,7 @@ trait Defaults
     use InteractsWithForms;
     use WithRateLimiting;
 
-    public function getUser(): FilamentUser
+    public function getUser(): User
     {
         $user = Filament::auth()->user();
 
