@@ -2,12 +2,16 @@
 
 namespace Stephenjude\FilamentTwoFactorAuthentication\Pages;
 
+use Filament\Actions\Action;
 use Filament\Facades\Filament;
+use Filament\Schemas\Components\Actions;
+use Filament\Schemas\Schema;
 use Illuminate\Contracts\Support\Htmlable;
+use Stephenjude\FilamentTwoFactorAuthentication\TwoFactorAuthenticationPlugin;
 
 class Setup extends BaseSimplePage
 {
-    protected static string $view = 'filament-two-factor-authentication::pages.setup';
+    protected string $view = 'filament-two-factor-authentication::pages.setup';
 
     public ?array $data = [];
 
@@ -21,5 +25,24 @@ class Setup extends BaseSimplePage
     public function getTitle(): string | Htmlable
     {
         return '';
+    }
+
+    public function utilityActionsForm(Schema $schema): Schema
+    {
+        return $schema
+            ->components([
+                Actions::make([
+                    Action::make('dashboard')
+                        ->visible(
+                            ! TwoFactorAuthenticationPlugin::get()->hasForcedTwoFactorSetup()
+                            || filament()->auth()->user()->hasEnabledTwoFactorAuthentication()
+                        )
+                        ->label(__('filament-two-factor-authentication::section.dashboard'))
+                        ->url(fn () => \filament()->getCurrentOrDefaultPanel()->getUrl())
+                        ->color('gray')
+                        ->icon('heroicon-o-home')
+                        ->link(),
+                ])->fullWidth(),
+            ]);
     }
 }
